controller/pyproject.toml
[project]
name = "weave-controller"
version = "0.1.0"
description = "Weave Control Plane using WireGuard and FRRouting"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.0.0",
    "argon2-cffi>=23.1.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "anyio>=4.0.0",
]

[tool.ruff]
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "require_preauth: tests that run with REQUIRE_PREAUTH=True",
]
